[REFACTOR]: apply classes tailwind in globalcss
