+++ components/global/Content.tsx
@@ -1,5 +1,5 @@
 import { ChildrinProps } from '@interfaces/interfaces';
 
 export default function Content({ children }: ChildrinProps) {
-  return <main className='px-10 py-5'>{children}</main>;
+  return <main className='size-box'>{children}</main>;
 }
