+++ components/global/Footer/index.tsx
@@ -3,7 +3,7 @@ import Shared from './Shared';
 
 const Footer = () => {
   return (
-    <footer className='px-10 py-5 text-gray-500 body-font'>
+    <footer className='size-box'>
       <div className='flex justify-center md:justify-between md:items-center'>
         <Copyright />
         <Shared />
