+++ components/global/NavBar/Menu.tsx
@@ -2,33 +2,35 @@ import { AppContext } from '@components/layout';
 import { MenuProps } from '@interfaces/interfaces';
 import { useContext } from 'react';
 import { FaBars, FaTimes } from 'react-icons/fa';
-import Links from '../Links';
+import Links from './Links';
 
-export default function Menu({ links }: MenuProps) {
+const Menu = ({ links }: MenuProps) => {
   const { nav, setNav } = useContext(AppContext);
   return (
-    <>
+    <nav>
       <div>
         <Links
           links={links}
           classList='hidden md:flex'
-          classItem='px-5 cursor-pointer font-medium text-gray-400 hover:scale-105 hover:text-gray-100'
+          classItem='item'
         />
       </div>
       <div
         onClick={() => setNav(!nav)}
-        className='cursor-pointer pr-5 z-10 text-gray-400 md:hidden'
+        className='cursor-pointer pr-5 z-10 md:hidden'
       >
         {nav ? <FaTimes size={30} /> : <FaBars size={30} />}
       </div>
       {nav && (
         <Links
           links={links}
-          classList='flex flex-col justify-center items-center absolute top-0 left-0 w-full h-screen bg-gradient-to-b from-black to-gray-800 text-gray-400'
-          classItem='px-4 cursor-pointer capitalize py-6 text-3xl'
+          classList='burger-menu'
+          classItem='item capitalize py-6 text-3xl'
           onClick={() => setNav(!nav)}
         />
       )}
-    </>
+    </nav>
   );
-}
+};
+
+export default Menu;
