+++ styles/globals.css
@@ -18,12 +18,16 @@ html {
   min-height: 100vh;
 }
 
-section {
-  @apply text-gray-400;
+.size-box {
+  @apply px-8 py-3 w-full text-gray-400;
 }
 
-.project1 {
-  background-image: url('/assets/project1.jpg');
+.item {
+  @apply cursor-pointer hover:text-gray-200 hover:scale-105 px-5;
+}
+
+.burger-menu {
+  @apply flex flex-col justify-center items-center absolute top-0 left-0 w-full h-screen bg-gradient-to-b from-black to-gray-800;
 }
 
 .name-of-developer {
